feat: refactor type definitions
