+++ src/components/features/Theme/ThemeHeroSection/index.tsx
@@ -1,9 +1,9 @@
 import styled from '@emotion/styled';
 import { useNavigate } from 'react-router-dom';
 
-import type { ThemeData } from '@/api/types/apiTypes';
 import { Container } from '@/components/common/layouts/Container';
 import { breakpoints } from '@/styles/variants';
+import type { ThemeData } from '@/types';
 
 type Props = {
   currentTheme: ThemeData | undefined;
