+++ src/types/mock.ts
@@ -7,6 +7,8 @@ export const ThemeMockData: ThemeData = {
   title: '예산은 가볍게, 감동은 무겁게❤️',
   description: '당신의 센스를 뽐내줄 부담 없는 선물',
   backgroundColor: '#4b4d50',
+  imageURL:
+    'https://st.kakaocdn.net/product/gift/gift_theme/20220216170226_38ba26d8eedf450683200d6730757204.png',
 };
 
 export const ThemeMockList = [ThemeMockData];
